LayoutTests/ietestcenter/Javascript/15.3.3.2-1-expected.txt

15.3.3.2-1 (TestCases/chapter15/15.3/15.3.3/15.3.3.2/15.3.3.2-1.js) Function.length - data property with value 1

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
